pyproject.toml
[tool.poetry]
name = "ppt-generator"
version = "0.1.0"
description = "api endpoint to create powerpoint from instructions"
authors = ["Reagan <rrr.chanhiulok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.44.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
fastapi = "^0.114.0"
python-multipart = "^0.0.9"
python-pptx = "^1.0.2"
pathvalidate = "^3.2.1"
ruff = "^0.6.4"
mypy = "^1.11.2"
replicate = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
